% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTopicsToSeuratObject.R
\name{addTopicsToSCE}
\alias{addTopicsToSCE}
\title{Adds topic information to SingleCellExperiment metadata}
\usage{
addTopicsToSCE(model, Object)
}
\arguments{
\item{model}{LDA model output}

\item{Object}{SCE object containing the data the model was created with.}
}
\value{
SingleCellExperiment object with updated metadata
}
\description{
This function adds the topic-document information as individual columns to the SingleCellExperiment metadata for each topic
}
\examples{
addTopicsToSCE(LDAmodel, SCE)

}
