% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\alias{generateBasicAuth}
\alias{setAuthentication}
\alias{getAuthentication}
\title{Generate Authorization Header}
\usage{
generateBasicAuth(username, password)

setAuthentication(string = Sys.getenv("TITAN_BASIC_AUTH"))

getAuthentication()
}
\arguments{
\item{username, password}{Credentials to encode.}

\item{string}{Base64 encoded string, as returned by
\code{\link[=generateBasicAuth]{generateBasicAuth()}}.}
}
\description{
Generate the basic authorization header for
authentication.
}
\details{
Use the output of \code{generateBasicAuth} as
input to \code{setAuthentication}. This makes it such
that the \verb{/metrics} endpoint requires the authentication
defined. Then use the credentials in the prometheus
configuration file.

basic_auth:
username: usr
password: pwd
}
