% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binned.R
\name{Binned}
\alias{Binned}
\title{Binned}
\description{
Superclass for binned metrics: \link{Histogram} and \link{Summary},
used for convenience and DRY, do not use directly.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Binned$new()}}
\item \href{#method-observe}{\code{Binned$observe()}}
\item \href{#method-clone}{\code{Binned$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binned$new(
  name,
  help,
  predicate,
  labels = NULL,
  unit = NULL,
  type = c("histogram", "summary")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the metric.}

\item{\code{help}}{Help text describing the metric.}

\item{\code{predicate}}{Predicate function to bin observations.}

\item{\code{labels}}{Character vector of labels available.}

\item{\code{unit}}{Unit of metric.}

\item{\code{type}}{Metric type.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-observe"></a>}}
\if{latex}{\out{\hypertarget{method-observe}{}}}
\subsection{Method \code{observe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binned$observe(val, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to observe and bin using the \code{predicate}.}

\item{\code{...}}{Key value pair of labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Observe
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Binned$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
