% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historgram.R
\name{bucket}
\alias{bucket}
\alias{is.bucket}
\title{bucket}
\usage{
bucket(label, value)

is.bucket(obj)
}
\arguments{
\item{label}{Name of the label.}

\item{value}{Value of the label.}

\item{obj}{Object to check.}
}
\value{
An object of class \code{bucket}.
}
\description{
Create a Bucket, needed for the \code{predicate} of
the \link{Histogram} which must return an object of
class \code{bucket}.
}
\examples{
bucket("0.3", 2)

pred <- function(x){
 if(x > .5)
   return(bucket("1", x))

 bucket(".5", x)
}

result <- pred(.8)
is.bucket(result)

}
