% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTITAN.R
\name{calcDistMat}
\alias{calcDistMat}
\title{Calculates input matrix}
\usage{
calcDistMat(Object, centroid_df, quant_cut, scale_factor, seed.number = 8)
}
\arguments{
\item{Object}{Seurat object containing gene expression information.}

\item{centroid_df}{Dataframe output by getCentroids. Inverse distance matrix is calculated on this matrix}

\item{quant_cut}{Quantile value for the number of centroid distance values to keep for each cell. Recommended range is between .75 to .95, but other values are viable.}

\item{scale_factor}{Integer value to scale down distance values by. Recommended range is between 3 to 15, but otherr values are viable}

\item{seed.number}{random integer to set seed}
}
\value{
dataframe
}
\description{
This function calculates the scaled inverse distance matrix and concatenates it with the gene matrix in preparation for inputting it into the spTITAN algorithm
}
\examples{
calcDistMat(Object, centroid_df)

}
