% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{Counter}
\alias{Counter}
\title{Counter}
\description{
Counter

Counter
}
\section{Super class}{
\code{\link[titan:Titan]{titan::Titan}} -> \code{Counter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Counter$new()}}
\item \href{#method-inc}{\code{Counter$inc()}}
\item \href{#method-set}{\code{Counter$set()}}
\item \href{#method-clone}{\code{Counter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Titan" data-id="counter">}\href{../../titan/html/Titan.html#method-counter}{\code{titan::Titan$counter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Titan" data-id="runApp">}\href{../../titan/html/Titan.html#method-runApp}{\code{titan::Titan$runApp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$new(name, help, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$inc(value = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$set(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
