% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScores.R
\name{GeneScores}
\alias{GeneScores}
\title{GeneScoreDistubiton}
\usage{
GeneScores(Model)
}
\arguments{
\item{Model}{LDA Model}
}
\value{
Vmatrix of Gene-Topic Scores
}
\description{
This functionreturns a matrix of genescores for each topic
}
\examples{
GeneScores(Model)

}
