% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTITAN.R
\name{getCentroids}
\alias{getCentroids}
\title{Finds centroids}
\usage{
getCentroids(Object, pos_df, k, seed.number = 8)
}
\arguments{
\item{Object}{Object containing the data the model was created with.}

\item{pos_df}{Dataframe containing spatial x and y coordinates of the cells. The rownames of the dataframe need to be the cell names, followed by the x coordinates in column1, then the y coordinates in column2.}

\item{k}{Number of centroids to use}

\item{seed.number}{random integer to set seed}
}
\value{
dataframe
}
\description{
This function finds centroids in the x,y space and plots them to visualize their location
}
\examples{
getCentroids(SeuratObj, cell_positions, k)

}
