% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{getMetrics}
\alias{getMetrics}
\title{Retrieve Metrics from Endpoint}
\usage{
getMetrics(endpoint, auth = NULL)
}
\arguments{
\item{endpoint}{URL to the endpoint.}

\item{auth}{Authentication string as returned
by \code{\link[=generateBasicAuth]{generateBasicAuth()}}.}
}
\description{
Retrieve the metrics via the endpoint,
useful for testing authentication.
}
