% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopics.R
\name{GetTopics}
\alias{GetTopics}
\title{Extract Topic Information}
\usage{
GetTopics(model, Object)
}
\arguments{
\item{model}{LDA model.}

\item{Object}{Seurat object containing the data the model was created with.}
}
\value{
document-topic matrix
}
\description{
This function extracts the topic information from the LDA model. The output matrix is the document-topic matrix showing how much each cell corresponds with each topic
}
\examples{
GetTopics(LDA_model, SeuratObj)

}
