% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Creates vector of annotation colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Desired length of vector}
}
\value{
Vector of colors for plotting
}
\description{
This function creates a vector of colors
}
\examples{
gg_color_hue(50)

}
