% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{heatmap}
\alias{heatmap}
\title{Topic Heatmap}
\usage{
heatmap(Object, topics, AnnoVector, AnnoName)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{topics}{document-topic matrix}

\item{AnnoVector}{Vector of cell annotations}

\item{AnnoName}{Name of cell annotation}
}
\value{
pheatmap object
}
\description{
This function produces a heatmap showing how much each cell lies within each topic
}
\examples{
heatmap(SeuratObj, DocTopMat, SeuratObj$annotation, "NameOfAnnotation")

}
