% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{heatmapClusterTopics}
\alias{heatmapClusterTopics}
\title{Clustered Topic Heatmap}
\usage{
heatmapClusterTopics(Object, topics, AnnoVector, AnnoName)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{topics}{document-topic matrix}

\item{AnnoVector}{Vector of cell annotations}

\item{AnnoName}{Name of cell annotation}
}
\value{
pheatmap object
}
\description{
This function produces a heatmap where the columns are clustered based on the expression pattern of the topics
}
\examples{
heatmap(SeuratObj, DocTopMat, SeuratObj$annotation, "NameOfAnnotation")

}
