% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{heatmapSortByTopicAsWellAsAnno}
\alias{heatmapSortByTopicAsWellAsAnno}
\title{Topic Heatmap - sorted by topic and annotation}
\usage{
heatmapSortByTopicAsWellAsAnno(Object, topics, sortByTopic = "Topic_16",
  AnnoVector, AnnoName)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{topics}{document-topic matrix}

\item{sortByTopic}{topic to be sorted by}

\item{AnnoVector}{Vector of cell annotations}

\item{AnnoName}{Name of cell annotation}
}
\value{
pheatmap object
}
\description{
This function produces a heatmap showing how much each cell lies within each topic. The cells in the heatmap are sorted based on their scores in a given topic. The annotations are also sorted by their values.
}
\examples{
heatmap(SeuratObj, DocTopMat, "Topic_16", SeuratObj$annotation, "NameOfAnnotation")

}
