% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{HeatmapTopic}
\alias{HeatmapTopic}
\title{Topic Heatmap}
\usage{
HeatmapTopic(Object, topics, AnnoVector, AnnoName, clusterTopics = F)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{topics}{document-topic matrix}

\item{AnnoVector}{Vector of cell annotations}

\item{AnnoName}{Name of cell annotation}

\item{clusterTopics}{Logical whether to cluster topics or not}
}
\value{
pheatmap object
}
\description{
This function produces a Heatmap showing how much each cell lies within each topic
}
\examples{
Heatmap(SeuratObj, DocTopMat, SeuratObj$annotation, "NameOfAnnotation")

}
