% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeAndAddTopics.R
\name{ImputeAndAddTopics}
\alias{ImputeAndAddTopics}
\title{Adds imputed Topics to a seruat object}
\usage{
ImputeAndAddTopics(
  Object,
  model,
  TopicPrefix = "ImputedTopic",
  assayName = "RNA"
)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{model}{LDA model output}

\item{TopicPrefix}{Prefix of Topics to be added to your metaData (Default "ImputedTopic")}

\item{assayName}{The name of the assay to associate with the reduction}
}
\value{
Seurat object with updated metadata of imputed geneExpression
}
\description{
This function adds the topic-document information as individual columns to the Seurat metadata for each topic
}
\examples{
ImputeAndAddTopics(SeuratObj, LDAmodel, TopicPrefix = "ImputedTopic")

}
