% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAelbowPlot.R
\name{LDAelbowPlot}
\alias{LDAelbowPlot}
\title{Topic Elbow Plot}
\usage{
LDAelbowPlot(model_dir, SO)
}
\arguments{
\item{model_dir}{Directory containing the models created using a varying number of topics.}

\item{SO}{Seurat object containing the data the model was created with.}
}
\value{
ggplot object
}
\description{
This function helps to determine the proper number of topics to use in your LDA model. Plots the rate of perplexity change versus the number of topic. The ideal topic number is the "elbow" of the plot.
}
\examples{
LDAelbowPlot(test_dir, SeuratObj)

}
