% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{Metric}
\alias{Metric}
\title{Metrics
Store Prometheus metrics}
\description{
Metrics
Store Prometheus metrics

Metrics
Store Prometheus metrics
}
\section{Super class}{
\code{\link[titan:Titan]{titan::Titan}} -> \code{Metric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Metric$new()}}
\item \href{#method-value}{\code{Metric$value()}}
\item \href{#method-retrieve}{\code{Metric$retrieve()}}
\item \href{#method-clone}{\code{Metric$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Titan" data-id="counter">}\href{../../titan/html/Titan.html#method-counter}{\code{titan::Titan$counter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Titan" data-id="runApp">}\href{../../titan/html/Titan.html#method-runApp}{\code{titan::Titan$runApp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$new(
  name,
  help,
  type = c("counter", "gauge", "histogram", "summary"),
  label = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$value(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
