% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ns}
\alias{ns}
\alias{setTitanNamespace}
\alias{getTitanNamespace}
\title{Set a Namespace}
\usage{
setTitanNamespace(name = NULL)

getTitanNamespace()
}
\arguments{
\item{name}{Namespace.}
}
\description{
Set a namespace, all metrics name will
be preceded by the namespace and an underscore:
\code{namespace_metricName}.
}
\note{
Rerun the function to remove the namespace.
}
