% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_runLDA.R
\name{parallel_runLDA}
\alias{parallel_runLDA}
\title{Runs a series of LDA models in parallel}
\usage{
parallel_runLDA(Object, ntopics, outdir, alpha = 50, beta = 0.1,
  varFeatures = 5000, iterations = 500, burnin = 250,
  top_start = 5, top_end = 50)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{ntopics}{Number of topics to be used in the model}

\item{outdir}{Directory for all of the output models}

\item{alpha}{the value for alpha in the LDA model}

\item{beta}{the value for beta in the LDA model}

\item{varFeatures}{the number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.}

\item{iterations}{the number of iterations used when learning the LDA model.}

\item{burnin}{number of iterations to run to allow the model to learn before calculating certain statistics. Models start at random points, so this allows model to get closer to the fit before certain statistics are calculated.}

\item{top_start}{Lowest number of topics to run a model with}

\item{top_end}{Highest number of topics to run a model with}
}
\value{
saves the different models to the given output directory
}
\description{
This function runs a series of LDA models on scRNA-seq expression data in parallel.
}
\examples{
parallel_runLDA(SeuratObj, ntopics = 20, "model_dir/")

}
