% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2Topics.R
\name{plot2Topics}
\alias{plot2Topics}
\title{Plot2Topics}
\usage{
plot2Topics(SeuratObj, model, topics, GOI, plot_split, IntIdents)
}
\arguments{
\item{SeuratObj}{Seurat Object containing the data the model was created with.}

\item{model}{Output of the LDA function}

\item{topics}{The two topics that will plotted along the axes}

\item{GOI}{The cells on the plot will be colored by how much they express the given gene. If "CC" is entered, the cells will by colored by their cell cycle phase. Cell cycle must be calculated prior.}

\item{plot_split}{A column in the metadata of the Seurat Object to split the plots by}

\item{IntIdents}{Specific factor levels in the plot_split to plot. (If left empty, all factor levels will be plotted)}
}
\value{
Prints plot
}
\description{
This function allows the user to choose two topics and plot them along the x and y axis of a plot
}
\examples{
Plot2Topics(SeuratObj = my_SO, model = my_LDA, topics = c(1,2), GOI = "CC")

}
