% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{prTitan}
\alias{prTitan}
\title{Plumber}
\usage{
prTitan(file = NULL, ..., latency = NULL)
}
\arguments{
\item{file}{Plumber file as passed to \code{\link[plumber:pr]{plumber::pr()}}.}

\item{...}{Any other argument to pass to \code{\link[plumber:pr]{plumber::pr()}}.}

\item{latency}{Pass a character string to use as metric name
to track request latency. Leave on \code{NULL} to not track.}
}
\description{
Serve a plumber API with metrics, use this function
like you would use \code{\link[plumber:pr]{plumber::pr()}}.
}
