% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{Registry}
\alias{Registry}
\title{Registry}
\description{
Interface to the main registry of metrics of titan,
should not be interacted with directly.
}
\author{
John Coene, \email{john@opifex.org}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Registry$new()}}
\item \href{#method-store}{\code{Registry$store()}}
\item \href{#method-get}{\code{Registry$get()}}
\item \href{#method-clone}{\code{Registry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Initialise a new storage for a metric.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$store(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{Object to store.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Store something in the registry.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Retrieve something from the registry.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
