% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reloadConfig}
\alias{reloadConfig}
\title{Reload the configuration}
\usage{
reloadConfig(prometheus = "http://localhost:9090")
}
\arguments{
\item{prometheus}{URL to the prometheus service, including
the port if necessary.}
}
\description{
Tell Prometheus to reload the configuration file.
Note that this only works if Prometheus was launched
with the \code{--web.enable-lifecycle} flag.
}
