% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{renderMetrics}
\alias{renderMetrics}
\title{Render All Metrics}
\usage{
renderMetrics()
}
\description{
Renders all metrics from the registry.
}
\seealso{
\code{\link[=previewMetrics]{previewMetrics()}} for a more readable
output.
}
