% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDA.R
\name{runLDA}
\alias{runLDA}
\title{Runs LDA Model}
\usage{
runLDA(Object, ntopics, alpha = 50, beta = 0.1, varFeatures = 5000,
  iterations = 500, burnin = 250)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{ntopics}{Number of topics to be used in the model}

\item{alpha}{the value for alpha in the LDA model}

\item{beta}{the value for beta in the LDA model}

\item{varFeatures}{the number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.}

\item{iterations}{the number of iterations used when learning the LDA model.}

\item{burnin}{number of iterations to run to allow the model to learn before calculating certain statistics. Models start at random points, so this allows model to get closer to the fit before certain statistics are calculated.}
}
\value{
LDA Model
}
\description{
This function runs an LDA model on scRNA-seq expression data
}
\examples{
runLDA(SeuratObj, ntopics = 20)

}
