% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCE_FeaturePlot.R, R/TopicWordDistribution .R
\name{SCE_FeaturePlot}
\alias{SCE_FeaturePlot}
\title{FeaturePlot function for SingleCellExperiment}
\usage{
SCE_FeaturePlot(Model, nGenes = 50)

SCE_FeaturePlot(Model, nGenes = 50)
}
\arguments{
\item{Model}{LDA Model Object}

\item{nGenes}{Number of genes do you want returned per topic}

\item{Object}{SingleCellExperiment object with topics added to it}

\item{pt.size}{Size of points on plot}

\item{topics}{The topics to plot on the feature plot}

\item{ncol}{Number of columns to use when faceting plots}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature}
}
\value{
Returns a ggplot object or list of ggplot objects

Returns a ggplot object or list of ggplot objects
}
\description{
Plots the Seurat FeaturePlot function for the topics within SingleCellExperiment objects

Returns a dataframe with the top n genes per topic
}
\examples{
SCE_FeaturePlot(SCE, 0.01, 1:12, 4, 'q5')

SCE_FeaturePlot(SCE, 0.01, 1:12, 4, 'q5')

}
