% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTITAN.R
\name{spTITANelbowPlot}
\alias{spTITANelbowPlot}
\title{spTITAN Topic Elbow Plot}
\usage{
spTITANelbowPlot(model_dir, input_df, varFeatures = 5000, assayName = "RNA")
}
\arguments{
\item{model_dir}{Directory containing the models created using a varying number of topics.}

\item{input_df}{Output of calcDistMat}

\item{assayName}{The name of the assay holding the source data}

\item{VarFeatures}{the number of variable features to use in the LDA model. MUST MATCH WITH MODELS IN model_dir}
}
\value{
ggplot object
}
\description{
This function helps to determine the proper number of topics to use in your LDA model. Plots the rate of perplexity change versus the number of topic. The ideal topic number is the "elbow" of the plot.
}
\examples{
spTITANelbowPlot(test_dir, SeuratObj)

}
