% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{Summary}
\alias{Summary}
\title{Summary}
\description{
Create a summary.
}
\author{
John Coene, \email{john@opifex.org}
}
\section{Super class}{
\code{\link[titan:Registry]{titan::Registry}} -> \code{Summary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Summary$new()}}
\item \href{#method-observe}{\code{Summary$observe()}}
\item \href{#method-clone}{\code{Summary$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="decValue">}\href{../../titan/html/Registry.html#method-decValue}{\code{titan::Registry$decValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="get">}\href{../../titan/html/Registry.html#method-get}{\code{titan::Registry$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="getValue">}\href{../../titan/html/Registry.html#method-getValue}{\code{titan::Registry$getValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="incValue">}\href{../../titan/html/Registry.html#method-incValue}{\code{titan::Registry$incValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="setValue">}\href{../../titan/html/Registry.html#method-setValue}{\code{titan::Registry$setValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="store">}\href{../../titan/html/Registry.html#method-store}{\code{titan::Registry$store()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Summary$new(name, help, predicate, labels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the summary.}

\item{\code{help}}{Help text describing the summary.}

\item{\code{predicate}}{A callback function that take a
single argument, a numeric vector of length 1 and
returns a \code{\link[=bucket]{bucket()}}.}

\item{\code{labels}}{Labels to use for the summary, a
character vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-observe"></a>}}
\if{latex}{\out{\hypertarget{method-observe}{}}}
\subsection{Method \code{observe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Summary$observe(val, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value observed.}

\item{\code{...}}{Labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Observe a value.

This runs the value through the \code{predicate} and
updates meta data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
