% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiorix.R
\name{titan}
\alias{titan}
\title{Ambiorix Middleware}
\usage{
titan(path = "metrics")
}
\arguments{
\item{path}{Path on which to serve the metrics.
Default recommended.}
}
\description{
Middleware for ambiorix.
}
