% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{titanApp}
\alias{titanApp}
\title{Serve Application with metrics}
\usage{
titanApp(ui, server, ...)
}
\arguments{
\item{ui, server}{UI definition and server function
as passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{...}{Any other arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Serve Application with metrics
}
