% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicWordTable.R
\name{TopicWordTable}
\alias{TopicWordTable}
\title{TopicWordTable function for SingleCellExperiment}
\usage{
TopicWordTable(Model, nGenes = 50)
}
\arguments{
\item{Model}{LDA Model Object}

\item{nGenes}{Number of genes do you want returned per topic}
}
\value{
Returns a dataframe with the top n genes per topic
}
\description{
Dataframe with the top n genes per topic
}
\examples{
SCE_FeaturePlot(Model = LDAmodel, nGenes = 50)

}
