% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferTopics.R
\name{TransferTopics}
\alias{TransferTopics}
\title{Transfers topics from one dataset to another}
\usage{
TransferTopics(Object, Model, TopicPrefix = "Topic", assayName = "RNA")
}
\arguments{
\item{Object}{Seurat object to transfer topics onto}

\item{Model}{LDA model using data from another dataset}

\item{TopicPrefix}{Common prefix for topics in the LDA model}

\item{assayName}{The name of the assay to associate with the reduction}
}
\value{
Seurat object with topics transferred into its metadata
}
\description{
This function takes the topics from one dataset and, using the word-topic distribution from that dataset, to impute these topics onto a new dataset.
}
\examples{
TransferTopics(NewSeuratObj, Old_model)

}
