% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTopicsToSeuratObject.R
\name{addTopicsToSeuratObject}
\alias{addTopicsToSeuratObject}
\title{Adds topic information to Seurat metadata}
\usage{
addTopicsToSeuratObject(model, Object, assayName = "RNA")
}
\arguments{
\item{model}{LDA model output}

\item{Object}{Seurat object containing the data the model was created with.}

\item{assayName}{The name of the assay to associate with the reduction}
}
\value{
Seurat object with updated metadata
}
\description{
This function adds the topic-document information as individual columns to the Seurat metadata for each topic
}
\examples{
addTopicsToSeuratObject(LDAmodel, SeuratObj)

}
