% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{bucket}
\alias{bucket}
\alias{is.bucket}
\title{Bucket}
\usage{
bucket(label, value)

is.bucket(obj)
}
\arguments{
\item{label}{Label of bucket.}

\item{value}{Value of the bucket.}

\item{obj}{Object to check.}
}
\description{
Create buckets as must be returned by the
\link{Histogram} and \link{Summary} predicates.
}
