% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{Counter}
\alias{Counter}
\title{Counter}
\description{
Counter for Prometheus. The value of a counter,
can only increase over time.
}
\examples{

## ------------------------------------------------
## Method `Counter$new`
## ------------------------------------------------

c <- Counter$new(
 "refresh_btn_total", 
 "Button that refreshes the data",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")

## ------------------------------------------------
## Method `Counter$inc`
## ------------------------------------------------

c <- Counter$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")

## ------------------------------------------------
## Method `Counter$set`
## ------------------------------------------------

c <- Counter$new(
 "refresh_btn_total", 
 "Loads data button count"
) 

c$set(1)
c$set(5)

# will throw warning: 3 < 5
\dontrun{c$set(3)}
}
\author{
John Coene, \email{john@opifex.org}
}
\section{Super class}{
\code{\link[titan:Registry]{titan::Registry}} -> \code{Counter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Counter$new()}}
\item \href{#method-inc}{\code{Counter$inc()}}
\item \href{#method-set}{\code{Counter$set()}}
\item \href{#method-print}{\code{Counter$print()}}
\item \href{#method-clone}{\code{Counter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="get">}\href{../../titan/html/Registry.html#method-get}{\code{titan::Registry$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="store">}\href{../../titan/html/Registry.html#method-store}{\code{titan::Registry$store()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$new(name, help, labels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the counter.}

\item{\code{help}}{The help text discribing the counter.}

\item{\code{labels}}{Labels to use for this counter, a
character vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise a counter
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Counter$new(
 "refresh_btn_total", 
 "Button that refreshes the data",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$inc(val = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to increase the counter by.}

\item{\code{...}}{Named label and value pair(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Increase the value of the counter.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Counter$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$set(val, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to set the counter to.}

\item{\code{...}}{Named label and value pair(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the counter to a certain value.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Counter$new(
 "refresh_btn_total", 
 "Loads data button count"
) 

c$set(1)
c$set(5)

# will throw warning: 3 < 5
\dontrun{c$set(3)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print the counter.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
