% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{Counter}
\alias{Counter}
\title{Counter}
\description{
Counter

Counter
}
\section{Super class}{
\code{\link[titan:Titan]{titan::Titan}} -> \code{Counter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Counter$new()}}
\item \href{#method-inc}{\code{Counter$inc()}}
\item \href{#method-set}{\code{Counter$set()}}
\item \href{#method-clone}{\code{Counter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Titan" data-id="counter">}\href{../../titan/html/Titan.html#method-counter}{\code{titan::Titan$counter()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$new(name, help, namespace = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the counter.}

\item{\code{help}}{Help text for the counter.}

\item{\code{namespace}}{Namespace.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a counter.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$inc(value = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\emph{Positive} value (integer or numeric) to increase the
counter by.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Increase the counter.


If a negative value is passed the method
throws a warning and does not log metric.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$set(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set the counter to. This
value must be greater than the current counter
value or a warning is thrown and the value is
not logged.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the counter to a specific value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
