% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{Gauge}
\alias{Gauge}
\title{Gauge}
\description{
Gauge for Prometheus. The value of a gauge,
can go up or down, as opposed to counters.
}
\examples{

## ------------------------------------------------
## Method `Gauge$new`
## ------------------------------------------------

c <- Gauge$new(
 "refresh_btn_total", 
 "Button that refreshes the data",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")

## ------------------------------------------------
## Method `Gauge$inc`
## ------------------------------------------------

c <- Gauge$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")

## ------------------------------------------------
## Method `Gauge$dec`
## ------------------------------------------------

c <- Gauge$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(5)
c$dec(1)
c$dec()

## ------------------------------------------------
## Method `Gauge$set`
## ------------------------------------------------

c <- Gauge$new(
 "visitors_total", 
 "Total visitors right now"
) 

c$set(9)
c$set(4)

# will throw warning: 3 < 5
\dontrun{c$set(3)}
}
\author{
John Coene, \email{john@opifex.org}
}
\section{Super class}{
\code{\link[titan:Registry]{titan::Registry}} -> \code{Gauge}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Gauge$new()}}
\item \href{#method-inc}{\code{Gauge$inc()}}
\item \href{#method-dec}{\code{Gauge$dec()}}
\item \href{#method-set}{\code{Gauge$set()}}
\item \href{#method-print}{\code{Gauge$print()}}
\item \href{#method-clone}{\code{Gauge$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="decValue">}\href{../../titan/html/Registry.html#method-decValue}{\code{titan::Registry$decValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="get">}\href{../../titan/html/Registry.html#method-get}{\code{titan::Registry$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="getValue">}\href{../../titan/html/Registry.html#method-getValue}{\code{titan::Registry$getValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="incValue">}\href{../../titan/html/Registry.html#method-incValue}{\code{titan::Registry$incValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="setValue">}\href{../../titan/html/Registry.html#method-setValue}{\code{titan::Registry$setValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Registry" data-id="store">}\href{../../titan/html/Registry.html#method-store}{\code{titan::Registry$store()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$new(name, help, labels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the gauge.}

\item{\code{help}}{The help text discribing the gauge.}

\item{\code{labels}}{Labels to use for this gauge, a
character vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise a
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Gauge$new(
 "refresh_btn_total", 
 "Button that refreshes the data",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$inc(val = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to increase the gauge by.}

\item{\code{...}}{Named label and value pair(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Increase the value of the gauge.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Gauge$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(1, module = "homepage")
c$inc(1, module = "tableview")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dec"></a>}}
\if{latex}{\out{\hypertarget{method-dec}{}}}
\subsection{Method \code{dec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$dec(val = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to decrease the gauge by.}

\item{\code{...}}{Named label and value pair(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Decrease the value of the gauge.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Gauge$new(
 "refresh_btn_total", 
 "Loads data button count",
 labels = c("module")
) 

c$inc(5)
c$dec(1)
c$dec()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$set(val, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Value to set the gauge to.}

\item{\code{...}}{Named label and value pair(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the gauge to a certain value.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{c <- Gauge$new(
 "visitors_total", 
 "Total visitors right now"
) 

c$set(9)
c$set(4)

# will throw warning: 3 < 5
\dontrun{c$set(3)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print the gauge.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gauge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
