% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicHeatmaps.R
\name{HeatmapSortByTopic}
\alias{HeatmapSortByTopic}
\title{Topic Heatmap - sorted by topic}
\usage{
HeatmapSortByTopic(
  Object,
  topics,
  sortByTopic = "Topic_1",
  AnnoVector,
  AnnoName
)
}
\arguments{
\item{Object}{Seurat object containing the data the model was created with.}

\item{topics}{document-topic matrix}

\item{sortByTopic}{topic to be sorted by}

\item{AnnoVector}{Vector of cell annotations}

\item{AnnoName}{Name of cell annotation}
}
\value{
pheatmap object
}
\description{
This function produces a Heatmap showing how much each cell lies within each topic. The cells in the Heatmap are sorted based on their scores in a given topic.
}
\examples{
HeatmapSortByTopic(SeuratObj, DocTopMat, "Topic_16", SeuratObj$annotation, "NameOfAnnotation")

}
