% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{Metric}
\alias{Metric}
\title{Metrics
Store Prometheus metrics}
\description{
Metrics
Store Prometheus metrics

Metrics
Store Prometheus metrics
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Metric$new()}}
\item \href{#method-label}{\code{Metric$label()}}
\item \href{#method-value}{\code{Metric$value()}}
\item \href{#method-retrieve}{\code{Metric$retrieve()}}
\item \href{#method-render}{\code{Metric$render()}}
\item \href{#method-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$new(
  name,
  help,
  type = c("counter", "gauge", "histogram", "summary"),
  namespace = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the metric.}

\item{\code{help}}{Help text describing the metric.}

\item{\code{type}}{Metric type.}

\item{\code{namespace}}{Namespace.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new metric
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\if{latex}{\out{\hypertarget{method-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$label(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Label name.}

\item{\code{value}}{Value of the label.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add a label to the metric
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set the metric to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the metric value
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$retrieve()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Retrieve the metrics.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$render()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Render the metric
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
