% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_runLDA.R
\name{parallel_runLDA}
\alias{parallel_runLDA}
\title{Runs a series of LDA models in parallel}
\usage{
parallel_runLDA(Object, outdir, top_start = 5, top_end = 50,
  step = 5, cores = 10, alpha = 50, beta = 0.1,
  varFeatures = 5000, iterations = 500, burnin = 250)
}
\arguments{
\item{Object}{Object containing the data the model was created with.}

\item{outdir}{Directory for all of the output models}

\item{top_start}{Lowest number of topics to run a model with}

\item{top_end}{Highest number of topics to run a model with}

\item{step}{The step to use between top_start and top_end. (e.g a step of 5 with top_start = 5 and top_end = 50 will run 10 instances of the model using up to 50 topics counting up by 5)}

\item{cores}{Number of cores available. The recommended value is top_end/step}

\item{alpha}{the value for alpha in the LDA model}

\item{beta}{the value for beta in the LDA model}

\item{varFeatures}{the number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.}

\item{iterations}{the number of iterations used when learning the LDA model.}

\item{burnin}{number of iterations to run to allow the model to learn before calculating certain statistics. Models start at random points, so this allows model to get closer to the fit before certain statistics are calculated.}
}
\value{
saves the different models to the given output directory
}
\description{
This function runs a series of LDA models on scRNA-seq expression data in parallel.
}
\examples{
parallel_runLDA(SeuratObj, "model_dir")

}
