% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDA.R
\name{runLDA}
\alias{runLDA}
\title{Runs LDA Model}
\usage{
runLDA(
  Object,
  ntopics,
  alpha = 50,
  beta = 0.1,
  varFeatures = 5000,
  iterations = 500,
  burnin = 250,
  seed.number = 8,
  parallel = F,
  outDir = NULL,
  cores = 1,
  normalizationMethod = "CLR",
  assayName = "RNA"
)
}
\arguments{
\item{Object}{Object containing the data the model was created with.}

\item{ntopics}{Number of topics to be used in the model. If parallel == TRUE, a vector of topics to run should be inputted}

\item{alpha}{the value for alpha in the LDA model}

\item{beta}{the value for beta in the LDA model}

\item{varFeatures}{the number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.}

\item{iterations}{the number of iterations used when learning the LDA model.}

\item{burnin}{number of iterations to run to allow the model to learn before calculating certain statistics. Models start at random points, so this allows model to get closer to the fit before certain statistics are calculated.}

\item{seed.number}{random integer to set seed}

\item{parallel}{if TRUE, will run multiple models in parallel. NOT AVAILABLE ON WINDOWS}

\item{outDir}{if parallel = TRUE, the output directory for the multiple models}

\item{cores}{Number of cores to use, only applicable if parallel = TRUE}

\item{normalizationMethod}{Normalization method used by Seurat NormalizeData. Options are CLR, LogNormalize and RC.}

\item{assayName}{The name of the assay holding the source data}
}
\value{
LDA Model
}
\description{
This function runs an LDA model on scRNA-seq expression data
}
\examples{
runLDA(SeuratObj, ntopics = 20)

}
