% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicWordDistribution.R
\name{SCE_FeaturePlot}
\alias{SCE_FeaturePlot}
\title{TopicWordTable function for SingleCellExperiment}
\usage{
SCE_FeaturePlot(Model, nGenes = 50)
}
\arguments{
\item{Model}{LDA Model Object}

\item{nGenes}{Number of genes do you want returned per topic}
}
\value{
Returns a ggplot object or list of ggplot objects
}
\description{
Returns a dataframe with the top n genes per topic
}
\examples{
SCE_FeaturePlot(SCE, 0.01, 1:12, 4, 'q5')

}
