% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.R
\name{Summary}
\alias{Summary}
\title{Summary}
\description{
Create a histogram to bin observed metrics. Values
observed at preprocessed with the \code{predicate} function
passed at instantiation. This \code{predicate} function,
\strong{must} return an object of class \code{bucket}, see
\code{\link[=bucket]{bucket()}}.
}
\details{
While the \link{Histogram} counts up the values in each
bucket, the \link{Summary} sums these up over time.

Initialise
}
\section{Super class}{
\code{\link[titan:Binned]{titan::Binned}} -> \code{Summary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Summary$new()}}
\item \href{#method-clone}{\code{Summary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="titan" data-topic="Binned" data-id="observe">}\href{../../titan/html/Binned.html#method-observe}{\code{titan::Binned$observe()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Summary$new(name, help, predicate, labels = NULL, unit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the metric.}

\item{\code{help}}{Help text describing the metric.}

\item{\code{predicate}}{Predicate function to bin observations.}

\item{\code{labels}}{Character vector of labels available.}

\item{\code{unit}}{Unit of metric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
