% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titan.R
\name{Titan}
\alias{Titan}
\title{Registry}
\description{
Registry

Registry
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Titan$new()}}
\item \href{#method-counter}{\code{Titan$counter()}}
\item \href{#method-runApp}{\code{Titan$runApp()}}
\item \href{#method-clone}{\code{Titan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$new(namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-counter"></a>}}
\if{latex}{\out{\hypertarget{method-counter}{}}}
\subsection{Method \code{counter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$counter(name, help, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runApp"></a>}}
\if{latex}{\out{\hypertarget{method-runApp}{}}}
\subsection{Method \code{runApp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$runApp(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
