% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titan.R
\name{Titan}
\alias{Titan}
\title{Registry}
\description{
Registry

Registry
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Titan$new()}}
\item \href{#method-counter}{\code{Titan$counter()}}
\item \href{#method-clone}{\code{Titan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$new(namespace = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{Namespace to prefix all metric
names from this registry.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise a registry
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-counter"></a>}}
\if{latex}{\out{\hypertarget{method-counter}{}}}
\subsection{Method \code{counter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$counter(name, help)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the counter, must be unique
and match the following regex
\verb{[a-zA-Z_:][a-zA-Z0-9_:]*}.}

\item{\code{help}}{Help text describing the counter.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a counter
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Titan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
