% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{titanApp}
\alias{titanApp}
\title{Titan App}
\usage{
titanApp(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Run a shiny application with titan.
}
\author{
John Coene, \email{john@opifex.org}
}
