% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{withTitan}
\alias{withTitan}
\title{With Titan}
\usage{
withTitan(app)
}
\arguments{
\item{app}{An object of class \code{shiny.appobj}
as returned by \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A modified version of the shiny application.
}
\description{
Run a shiny application with titan.
}
\author{
John Coene, \email{john@opifex.org}
}
